Fix standings after first race in the season

+++ F1WM/ApiModel/Standings/ConstructorPositionAfterRace.cs
@@ -2,7 +2,7 @@ namespace F1WM.ApiModel
 {
 	public class ConstructorPositionAfterRace : ConstructorPosition
 	{
-		public float Change { get; set; }
+		public int Change { get; set; }
 		public float NotCountedTowardsChampionshipPoints { get; set; }
 	}
 }
